Create Release2
This is the updated code with all the correct Release 2 additions to the game. This includes the ability to pick the desired symbol for P1. This also includes added JLabels for the buttons to help make the game more user friendly. Along with this the new code includes the option to play against a much more challenging AI that predicts where you will play. This AI difficulty can be toggled easily between a Hard and Easy mode.